move : modal 공통 컴포넌트 폴더 생성 및 파일 이동
